+++ middlewares/jwt_bearer.py
@@ -1,5 +1,5 @@
 from fastapi import HTTPException, Request
-from jwt_manager import validate_token
+from utils.jwt_manager import validate_token
 from fastapi.security import HTTPBearer
 
 class JWTBearer(HTTPBearer):
